refactor: separate auth routes
